Change acted_at_turn field from FloatField to IntegerField
Updated the acted_at_turn field in the ActionHistory model to use IntegerField instead of FloatField. This change ensures data consistency and aligns the field type with its intended use for storing integer values. A migration file was added to apply this modification.

+++ ai_agent/models.py
@@ -104,7 +104,7 @@ class ActionTimeline(models.Model):
 
 class ActionHistory(models.Model):
     entity = models.ForeignKey(Entity, on_delete=models.CASCADE)
-    acted_at_turn = models.FloatField()
+    acted_at_turn = models.IntegerField()
     done = models.BooleanField(default=False)
     created_at = models.DateTimeField(auto_now_add=True)
